+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/base-fields.mdx
@@ -1,6 +1,6 @@
 ---
 title: 'Base fields'
-order: 1
+order: 2
 breadcrumb:
   - text: Forms
     href: /uilib/extensions/forms/
