+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/create-component.mdx
@@ -1,6 +1,6 @@
 ---
 title: 'Create your own component'
-order: 4
+order: 5
 breadcrumb:
   - text: Forms
     href: /uilib/extensions/forms/
