+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/extended-features.mdx
@@ -1,6 +1,6 @@
 ---
 title: 'Extended features'
-order: 3
+order: 4
 breadcrumb:
   - text: Forms
     href: /uilib/extensions/forms/
@@ -53,6 +53,10 @@ render(
   </StepsLayout>,
 )
 ```
+
+## [Value](/uilib/extensions/forms/extended-features/Value/)
+
+Value components are in practice like a field component whose value cannot be changed. They can be displayed with or without a label, and they format the value based on given rules. They can also connect to the surrounding [DataContext](/uilib/extensions/forms/extended-features/DataContext/).
 
 ## [Iterate](/uilib/extensions/forms/extended-features/Iterate/)
