+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/extended-features/Value.mdx
@@ -1,15 +1,15 @@
 ---
 title: 'Value'
 description: '`Value` components can be used to summarize any kind of data.'
-componentType: 'basis-api'
+componentType: 'advanced-api'
 hideInMenu: true
 breadcrumb:
   - text: Forms
     href: /uilib/extensions/forms/
-  - text: Create your component
-    href: /uilib/extensions/forms/create-component/
+  - text: Extended features
+    href: /uilib/extensions/forms/extended-features/
   - text: Value
-    href: /uilib/extensions/forms/create-component/Value/
+    href: /uilib/extensions/forms/extended-features/Value/
 ---
 
 import ListBaseValueComponents from './Value/ListBaseValueComponents'
