+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/extended-features/Value/BankAccountNumber.mdx
@@ -14,16 +14,16 @@ tabs:
 breadcrumb:
   - text: Forms
     href: /uilib/extensions/forms/
-  - text: Create your component
-    href: /uilib/extensions/forms/create-component/
+  - text: Extended features
+    href: /uilib/extensions/forms/extended-features/
   - text: Value
-    href: /uilib/extensions/forms/create-component/Value/
+    href: /uilib/extensions/forms/extended-features/Value/
   - text: BankAccountNumber
-    href: /uilib/extensions/forms/create-component/BankAccountNumber/
+    href: /uilib/extensions/forms/extended-features/BankAccountNumber/
 ---
 
-import Info from 'Docs/uilib/extensions/forms/create-component/Value/BankAccountNumber/info'
-import Demos from 'Docs/uilib/extensions/forms/create-component/Value/BankAccountNumber/demos'
+import Info from 'Docs/uilib/extensions/forms/extended-features/Value/BankAccountNumber/info'
+import Demos from 'Docs/uilib/extensions/forms/extended-features/Value/BankAccountNumber/demos'
 
 <Info />
 <Demos />
