+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/extended-features/Value/Boolean.mdx
@@ -14,16 +14,16 @@ tabs:
 breadcrumb:
   - text: Forms
     href: /uilib/extensions/forms/
-  - text: Create your component
-    href: /uilib/extensions/forms/create-component/
+  - text: Extended features
+    href: /uilib/extensions/forms/extended-features/
   - text: Value
-    href: /uilib/extensions/forms/create-component/Value/
+    href: /uilib/extensions/forms/extended-features/Value/
   - text: Boolean
-    href: /uilib/extensions/forms/create-component/Boolean/
+    href: /uilib/extensions/forms/extended-features/Boolean/
 ---
 
-import Info from 'Docs/uilib/extensions/forms/create-component/Value/Boolean/info'
-import Demos from 'Docs/uilib/extensions/forms/create-component/Value/Boolean/demos'
+import Info from 'Docs/uilib/extensions/forms/extended-features/Value/Boolean/info'
+import Demos from 'Docs/uilib/extensions/forms/extended-features/Value/Boolean/demos'
 
 <Info />
 <Demos />
