+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/extended-features/Value/Boolean/properties.mdx
