+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/extended-features/Value/Currency.mdx
@@ -14,16 +14,16 @@ tabs:
 breadcrumb:
   - text: Forms
     href: /uilib/extensions/forms/
-  - text: Create your component
-    href: /uilib/extensions/forms/create-component/
+  - text: Extended features
+    href: /uilib/extensions/forms/extended-features/
   - text: Value
-    href: /uilib/extensions/forms/create-component/Value/
+    href: /uilib/extensions/forms/extended-features/Value/
   - text: Currency
-    href: /uilib/extensions/forms/create-component/Currency/
+    href: /uilib/extensions/forms/extended-features/Currency/
 ---
 
-import Info from 'Docs/uilib/extensions/forms/create-component/Value/Currency/info'
-import Demos from 'Docs/uilib/extensions/forms/create-component/Value/Currency/demos'
+import Info from 'Docs/uilib/extensions/forms/extended-features/Value/Currency/info'
+import Demos from 'Docs/uilib/extensions/forms/extended-features/Value/Currency/demos'
 
 <Info />
 <Demos />
