+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/extended-features/Value/Date.mdx
@@ -14,16 +14,16 @@ tabs:
 breadcrumb:
   - text: Forms
     href: /uilib/extensions/forms/
-  - text: Create your component
-    href: /uilib/extensions/forms/create-component/
+  - text: Extended features
+    href: /uilib/extensions/forms/extended-features/
   - text: Value
-    href: /uilib/extensions/forms/create-component/Value/
+    href: /uilib/extensions/forms/extended-features/Value/
   - text: Date
-    href: /uilib/extensions/forms/create-component/Date/
+    href: /uilib/extensions/forms/extended-features/Date/
 ---
 
-import Info from 'Docs/uilib/extensions/forms/create-component/Value/Date/info'
-import Demos from 'Docs/uilib/extensions/forms/create-component/Value/Date/demos'
+import Info from 'Docs/uilib/extensions/forms/extended-features/Value/Date/info'
+import Demos from 'Docs/uilib/extensions/forms/extended-features/Value/Date/demos'
 
 <Info />
 <Demos />
