+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/extended-features/Value/Email.mdx
@@ -14,16 +14,16 @@ tabs:
 breadcrumb:
   - text: Forms
     href: /uilib/extensions/forms/
-  - text: Create your component
-    href: /uilib/extensions/forms/create-component/
+  - text: Extended features
+    href: /uilib/extensions/forms/extended-features/
   - text: Value
-    href: /uilib/extensions/forms/create-component/Value/
+    href: /uilib/extensions/forms/extended-features/Value/
   - text: Email
-    href: /uilib/extensions/forms/create-component/Email/
+    href: /uilib/extensions/forms/extended-features/Email/
 ---
 
-import Info from 'Docs/uilib/extensions/forms/create-component/Value/Email/info'
-import Demos from 'Docs/uilib/extensions/forms/create-component/Value/Email/demos'
+import Info from 'Docs/uilib/extensions/forms/extended-features/Value/Email/info'
+import Demos from 'Docs/uilib/extensions/forms/extended-features/Value/Email/demos'
 
 <Info />
 <Demos />
