+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/extended-features/Value/NationalIdentityNumber.mdx
@@ -14,16 +14,16 @@ tabs:
 breadcrumb:
   - text: Forms
     href: /uilib/extensions/forms/
-  - text: Create your component
-    href: /uilib/extensions/forms/create-component/
+  - text: Extended features
+    href: /uilib/extensions/forms/extended-features/
   - text: Value
-    href: /uilib/extensions/forms/create-component/Value/
+    href: /uilib/extensions/forms/extended-features/Value/
   - text: NationalIdentityNumber
-    href: /uilib/extensions/forms/create-component/NationalIdentityNumber/
+    href: /uilib/extensions/forms/extended-features/NationalIdentityNumber/
 ---
 
-import Info from 'Docs/uilib/extensions/forms/create-component/Value/NationalIdentityNumber/info'
-import Demos from 'Docs/uilib/extensions/forms/create-component/Value/NationalIdentityNumber/demos'
+import Info from 'Docs/uilib/extensions/forms/extended-features/Value/NationalIdentityNumber/info'
+import Demos from 'Docs/uilib/extensions/forms/extended-features/Value/NationalIdentityNumber/demos'
 
 <Info />
 <Demos />
