+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/extended-features/Value/Number.mdx
@@ -15,16 +15,16 @@ tabs:
 breadcrumb:
   - text: Forms
     href: /uilib/extensions/forms/
-  - text: Create your component
-    href: /uilib/extensions/forms/create-component/
+  - text: Extended features
+    href: /uilib/extensions/forms/extended-features/
   - text: Value
-    href: /uilib/extensions/forms/create-component/Value/
+    href: /uilib/extensions/forms/extended-features/Value/
   - text: Number
-    href: /uilib/extensions/forms/create-component/Number/
+    href: /uilib/extensions/forms/extended-features/Number/
 ---
 
-import Info from 'Docs/uilib/extensions/forms/create-component/Value/Number/info'
-import Demos from 'Docs/uilib/extensions/forms/create-component/Value/Number/demos'
+import Info from 'Docs/uilib/extensions/forms/extended-features/Value/Number/info'
+import Demos from 'Docs/uilib/extensions/forms/extended-features/Value/Number/demos'
 
 <Info />
 <Demos />
