+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/extended-features/Value/Number/info.mdx
