+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/extended-features/Value/PhoneNumber.mdx
@@ -14,16 +14,16 @@ tabs:
 breadcrumb:
   - text: Forms
     href: /uilib/extensions/forms/
-  - text: Create your component
-    href: /uilib/extensions/forms/create-component/
+  - text: Extended features
+    href: /uilib/extensions/forms/extended-features/
   - text: Value
-    href: /uilib/extensions/forms/create-component/Value/
+    href: /uilib/extensions/forms/extended-features/Value/
   - text: PhoneNumber
-    href: /uilib/extensions/forms/create-component/PhoneNumber/
+    href: /uilib/extensions/forms/extended-features/PhoneNumber/
 ---
 
-import Info from 'Docs/uilib/extensions/forms/create-component/Value/PhoneNumber/info'
-import Demos from 'Docs/uilib/extensions/forms/create-component/Value/PhoneNumber/demos'
+import Info from 'Docs/uilib/extensions/forms/extended-features/Value/PhoneNumber/info'
+import Demos from 'Docs/uilib/extensions/forms/extended-features/Value/PhoneNumber/demos'
 
 <Info />
 <Demos />
