+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/extended-features/Value/String.mdx
@@ -15,16 +15,16 @@ tabs:
 breadcrumb:
   - text: Forms
     href: /uilib/extensions/forms/
-  - text: Create your component
-    href: /uilib/extensions/forms/create-component/
+  - text: Extended features
+    href: /uilib/extensions/forms/extended-features/
   - text: Value
-    href: /uilib/extensions/forms/create-component/Value/
+    href: /uilib/extensions/forms/extended-features/Value/
   - text: String
-    href: /uilib/extensions/forms/create-component/String/
+    href: /uilib/extensions/forms/extended-features/String/
 ---
 
-import Info from 'Docs/uilib/extensions/forms/create-component/Value/String/info'
-import Demos from 'Docs/uilib/extensions/forms/create-component/Value/String/demos'
+import Info from 'Docs/uilib/extensions/forms/extended-features/Value/String/info'
+import Demos from 'Docs/uilib/extensions/forms/extended-features/Value/String/demos'
 
 <Info />
 <Demos />
