+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/feature-fields.mdx
@@ -1,6 +1,6 @@
 ---
 title: 'Feature fields'
-order: 2
+order: 3
 breadcrumb:
   - text: Forms
     href: /uilib/extensions/forms/
