+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/fields.mdx
@@ -1,24 +1,16 @@
 ---
-title: 'Field'
-description: '`Field` for interactive data driven components.'
-componentType: 'basis-api'
-hideInMenu: true
-showTabs: false
+title: 'Fields'
+order: 1
 breadcrumb:
   - text: Forms
     href: /uilib/extensions/forms/
-  - text: Create your component
-    href: /uilib/extensions/forms/create-component/
-  - text: Field
-    href: /uilib/extensions/forms/create-component/Field/
+  - text: Fields
+    href: /uilib/extensions/forms/fields/
 ---
 
-import ListBaseFieldComponents from '../base-fields/ListBaseFieldComponents'
-import ListFeatureFieldComponents from '../feature-fields/ListFeatureFieldComponents'
-
-import ListBaseInputComponents from './Field/ListBaseInputComponents'
-import ListBaseSelectionComponents from './Field/ListBaseSelectionComponents'
-import ListBaseToggleComponents from './Field/ListBaseToggleComponents'
+import ListBaseInputComponents from './fields/ListBaseInputComponents'
+import ListBaseSelectionComponents from './fields/ListBaseSelectionComponents'
+import ListBaseToggleComponents from './fields/ListBaseToggleComponents'
 
 # Field
 
@@ -49,13 +41,9 @@ The base components exists of several kinds, such as the:
 - **toggle category** – which contains of field types that allow the user to toggle between two values, such as `true` and `false`:
   <ListBaseToggleComponents />
 
-## All base components
+## Feature components
 
-<ListBaseFieldComponents />
-
-## Feature fields
-
-<ListFeatureFieldComponents />
+[Feature fields](/uilib/extensions/forms/feature-fields/) build on top of the base components to add more specific features targeting a given type of data.
 
 ## More info
